Add checkbutton to toggle CD-mark for all markers.
If the user has serveral (>10) markers, of which all or most should be
exported. Then all the markers have to be ticked in the editor list view
individually. This can be quite annoying. This branch adds check buttons
above the marker list to toggle CD, hidden, locked and glued for all
location markers.