Delay signal Session::Located in sync with the Audio engine
When we are syncing to an audio engine like jack transport, it might happen
that the current frame of the audio engine (AudioEngine::transport_frame()) is
not in sync with our current frame (Session::_transport_frame). If a receiver
of Session::Located() (e.g. Editor::located()) then asks for
Session::audible_frame() it gets the obsolete AudioEngine::transport_frame().

This causes the playhead to end up in an obsolete position.

The issue is described in Issue number 6887.

This commit is more gluing tape onto the issue rather than actually fix it. It
just delays the emission of Session::Located by 50 milliseconds.