Save action scripts and hook scripts into templates.
By now session scripts are the only lua scripts that make it into
templates. That's because saving session templates and creating a new session
from a template is done by libardour. Unfortunately libardour does not know
about action scripts and hook scripts as these are in the ui's regime.

Solution:

* Saving templates
  The current LuaInstance action script and hook script states are put into one
  XMLNode("LuaScripts"). A pointer to this node is passed to
  Session::save_template() as an additional parameter, so that it can be put into
  the template file.

* Creating new sessions from a template
  ARDOUR_UI::load_session() reads the template file again after the session has
  been created, if a template has been used. It reads the child "LuaScripts"
  and passes it to LuaInstance::set_state().